Update main.tex
In the criteria both $q$ and $\dot{u}$ should not be equal to zero in a point different from zero. The negation of this criteria would be $q \neq 0$ OR $\dot{q} \neq 0$

By the description in this line we have the same condition as on previous slide where the system is stable but not asymptotically.